% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.crop.R
\name{AFM.crop}
\alias{AFM.crop}
\title{Crops an AFM image}
\usage{
AFM.crop(obj, x0 = 1, y0 = 1, width.pixels, height.pixels, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{x0}{x origin location}

\item{y0}{y origin location}

\item{width.pixels}{width of cropped image in pixels}

\item{height.pixels}{height of cropped image in pixels}

\item{verbose}{output fitting parameters}
}
\value{
returns cropped image
}
\description{
Crops an AFM image
}
\author{
Thomas Gredig
}
