% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.dataType}
\alias{AFM.dataType}
\title{returns type of AFM image}
\usage{
AFM.dataType(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
string with AFM type, "image", "force", "frequency", "spectroscopy"
}
\description{
returns type of AFM image
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw')[1])
AFM.dataType(d)
}
\author{
Thomas Gredig
}
