% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.getLine.R
\name{AFM.getLine}
\alias{AFM.getLine}
\title{AFM line}
\usage{
AFM.getLine(obj, yPixel = 1, dataOnly = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{yPixel}{line number from 1 to max yPixel}

\item{dataOnly}{if \code{TRUE}, returns data instead of AFMdata object}

\item{verbose}{if \code{TRUE}, output additional information}
}
\value{
AFMdata object with line data or data frame with data
}
\description{
create a profile data line for a particular
line, given in pixels; easy way to reiterate
tbrough all lines for example
}
\examples{
filename = AFM.getSampleImages()[1]
afmd = AFM.import(filename)
afmd2 = AFM.getLine(afmd, 50)
plot(afmd2, addLines = TRUE)
head(AFM.linePlot(afmd2, dataOnly = TRUE))
}
\author{
Thomas Gredig
}
