% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.info.R
\name{AFM.info}
\alias{AFM.info}
\title{find AFM image information}
\usage{
AFM.info(filename, no = 1)
}
\arguments{
\item{filename}{filename of (Veeco, Park, AR, NanoSurf) AFM image including path}

\item{no}{channel number (for Veeco, NanoSurf, AR)}
}
\value{
returns gerenal information about the AFM image
}
\description{
find AFM image information
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
d = AFM.info(filename)
}
