% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.isImage}
\alias{AFM.isImage}
\title{checks if the object is an AFM image}
\usage{
AFM.isImage(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
\code{TRUE} if object is an AFM image
}
\description{
checks if the object is an AFM image
}
\examples{
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
AFM.isImage(d)
}
\author{
thomasgredig
}
