% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.lineProfile.R
\name{AFM.lineProfile}
\alias{AFM.lineProfile}
\title{Line Profile}
\usage{
AFM.lineProfile(obj, x1, y1, x2, y2, verbose = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{x1}{start x position in AFM units from bottom right}

\item{y1}{start y position in AFM units from bottom right}

\item{x2}{end x position in AFM units from bottom right}

\item{y2}{end y position in AFM units from bottom right}

\item{verbose}{if \code{TRUE}, output additional information}
}
\value{
AFMdata object with line data
}
\description{
create a profile data line across an image (d), providing
  the starting point (x1,y1) and end point (x2,y2). The start and end
  points are provided in length units of the images (such as micrometers)
}
\author{
Thomas Gredig
}
