% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.read.R
\name{AFM.read}
\alias{AFM.read}
\title{loads AFM image file}
\usage{
AFM.read(filename, no = 1)
}
\arguments{
\item{filename}{filename of (Veeco, Park, AR, NanoSurf) AFM image including path}

\item{no}{channel number (for Veeco, NanoSurf, AR)}
}
\value{
AFM image with attributes
}
\description{
loads AFM image file
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
d = AFM.read(filename)
}
