% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.info.R
\name{AFMinfo.item}
\alias{AFMinfo.item}
\title{returns an item from the AFMinfo object}
\usage{
AFMinfo.item(obj, itemName)
}
\arguments{
\item{itemName}{name to retrieve (ScanRate, ScanAngle, ...)}

\item{AFMinfo}{object}
}
\value{
string or number
}
\description{
returns an item from the AFMinfo object
}
\author{
Thomas Gredig
}
