% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMinfo.R
\name{AFMinfo}
\alias{AFMinfo}
\title{AFMinfo object}
\usage{
AFMinfo(filename)
}
\arguments{
\item{filename}{full filename of instrument file}
}
\value{
AFMinfo class object
}
\description{
heater information about the AFM image, includes scan rate, widths, etc.
}
\examples{
filename = system.file("extdata","Veeco_20160622.003",package="nanoscopeAFM")
h = AFMinfo(filename)
summary(h)
}
\author{
Thomas Gredig
}
