% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.math.R
\name{AFMmath.params}
\alias{AFMmath.params}
\title{Various computed AFM image parameters}
\usage{
AFMmath.params(obj)
}
\arguments{
\item{obj}{AFMdata class}
}
\value{
structure with various computed AFM image parameters
}
\description{
Various computed AFM image parameters
}
\examples{
filename = system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM")
d = AFM.math.params(AFM.import(filename))
}
