% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.NID_file.R
\name{check.NID_file}
\alias{check.NID_file}
\title{checks AFM NID file length}
\usage{
check.NID_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
true or false
}
\description{
checks AFM NID file length
}
\examples{
d = check.NID_file('example.nid')
}
