% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{.getChannelName}
\alias{.getChannelName}
\title{\code{.getChannelName} returns channel name based on header info}
\usage{
.getChannelName(h1, no)
}
\arguments{
\item{h1}{filename including path}

\item{no}{image number}
}
\value{
list of channels and additional information
}
\description{
\code{.getChannelName} returns channel name based on header info
}
