% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.NID_matrix.R
\name{flatten.NID_matrix}
\alias{flatten.NID_matrix}
\title{flattens a NID AFM image using a plane fit}
\usage{
flatten.NID_matrix(m)
}
\arguments{
\item{m}{matrix with AFM image}
}
\value{
flattened matrix with AFM image
}
\description{
flattens a NID AFM image using a plane fit
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
d = read.NID_file(filename)
m = matrix(d[[1]],nrow=256, ncol=256)
flatten.NID_matrix(m)
}
