% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{initialize,AFMdata-method}
\alias{initialize,AFMdata-method}
\title{Constructor method of AFMImage Class.}
\usage{
\S4method{initialize}{AFMdata}(
  .Object,
  data,
  x.conv,
  y.conv,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  history = character(0),
  description = character(0),
  fullfilename
)
}
\arguments{
\item{.Object}{an AFMdata object}
}
\description{
Constructor method of AFMImage Class.
}
\section{Slots}{

\describe{
\item{\code{data}}{list with objects ($z is a list with images)}

\item{\code{x.conv}}{conversion factor from pixels to nm}

\item{\code{y.conv}}{conversion factor from pixels to nm}

\item{\code{x.pixels}}{number of pixels in x-direction}

\item{\code{y.pixels}}{number of pixels in y-direction}

\item{\code{z.conv}}{(not used)}

\item{\code{z.units}}{vector with units for $z (deg, m)}

\item{\code{channel}}{vector with names of channels}

\item{\code{instrument}}{name of instrument (Park, Cypher, NanoSurf, Veeco)}

\item{\code{history}}{history of file changes}

\item{\code{description}}{AFM image description or note}

\item{\code{fullfilename}}{name of file}
}}

