% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{NID.checkFile}
\alias{NID.checkFile}
\title{checks AFM NID file length}
\usage{
NID.checkFile(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
mismatch in image size + header with file size (should be 0)
}
\description{
checks AFM NID file length
}
