% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.NIDchannel.Scale.R
\name{NID.getChannelScale}
\alias{NID.getChannelScale}
\title{returns the scales of a particular channel / image}
\usage{
NID.getChannelScale(headerList, imageNo = 1)
}
\arguments{
\item{headerList}{header list as obtained from read.NID_headerItems}

\item{imageNo}{1,2,3,4 denoting the number of the image}
}
\value{
data.frame with scales for (x,y,z)
}
\description{
returns the scales of a particular channel / image
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
h=read.NID_headerItems(filename)
NID.getChannelScale(h,1)
}
