% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NID.lineProfile.R
\name{NID.lineProfile}
\alias{NID.lineProfile}
\title{Line Profile}
\usage{
NID.lineProfile(d, x1, y1, x2, y2)
}
\arguments{
\item{d}{AFM image in data frame}

\item{x1}{start x position in AFM units from bottom right}

\item{y1}{start y position in AFM units from bottom right}

\item{x2}{end x position in AFM units from bottom right}

\item{y2}{end y position in AFM units from bottom right}
}
\value{
vector with line profile for data frame
}
\description{
create a profile data line across an image (d), providing
  the starting point (x1,y1) and end point (x2,y2). The start and end
  points are provided in length units of the images (such as micrometers)
}
