% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NID.loadImage.R
\name{NID.loadImage}
\alias{NID.loadImage}
\title{returns the scaled image + flattend image}
\usage{
NID.loadImage(filename, imageNo)
}
\arguments{
\item{filename}{filename including path}

\item{imageNo}{a number of the image, 1,2,3,4}
}
\value{
data.frame with four components, x and y-axes
        and the z-axis (original) and flattened
}
\description{
returns the scaled image + flattend image
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
d = NID.loadImage(filename)
ggplot(d, aes(x,y,fill=z.flatten)) + geom_raster()
}
