% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{print.AFMdata}
\alias{print.AFMdata}
\title{print essential information about the AFMdata object}
\usage{
\method{print}{AFMdata}(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
text with object information
}
\description{
print essential information about the AFMdata object
}
\examples{
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
print(d)
}
\author{
thomasgredig
}
