% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.afm.NID.R
\name{read.afm.NID}
\alias{read.afm.NID}
\title{loads header of AFM NID file}
\usage{
read.afm.NID(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with header
}
\description{
loads header of AFM NID file
}
\examples{
d = read.afm.NID('example.nid')
}
