% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.functions.R
\name{read.AFM_file}
\alias{read.AFM_file}
\title{DEPRECATED: use AFM.read()
loads AFM image file}
\usage{
read.AFM_file(filename, no = 1)
}
\arguments{
\item{filename}{filename of (Veeco, Park, AR, NanoSurf) AFM image including path}

\item{no}{channel number (for Veeco, NanoSurf, AR)}
}
\value{
AFM image with attributes
}
\description{
DEPRECATED: use AFM.read()
loads AFM image file
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
d = read.AFM_file(filename)
}
