% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_channelNames}
\alias{read.AR_channelNames}
\title{returns names of AR channels}
\usage{
read.AR_channelNames(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
string with channel names ("Height", "Amplitude")
}
\description{
returns names of AR channels
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
s1 = read.AR_channelNames(filename)
}
