% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_eofHeader.V2}
\alias{read.AR_eofHeader.V2}
\title{returns the version 2 header information tagged
at the end of the AR file, which includes several tags
including the $DataTypeList and $NumberOfFiles
ex: "HeightRetrace,AmplitudeRetrace,PhaseRetrace,ZSensorRetrace,"}
\usage{
read.AR_eofHeader.V2(wavefile, Verbose = FALSE)
}
\arguments{
\item{wavefile}{filename including path}

\item{Verbose}{if true, returns additional information}
}
\value{
list of channels and additional information
}
\description{
returns the version 2 header information tagged
at the end of the AR file, which includes several tags
including the $DataTypeList and $NumberOfFiles
ex: "HeightRetrace,AmplitudeRetrace,PhaseRetrace,ZSensorRetrace,"
}
\examples{
filename = system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM")
afmV2header = read.AR_eofHeader.V2(filename)
afmV2header$DataTypeList
}
