% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_file}
\alias{read.AR_file}
\title{loads Asylum Research Igor AFM wavefile
which can contain multiple images, returns channel
name and units in attributes}
\usage{
read.AR_file(filename, no = 1)
}
\arguments{
\item{filename}{filename including path}

\item{no}{number of the channel}
}
\value{
image with attributes
}
\description{
loads Asylum Research Igor AFM wavefile
which can contain multiple images, returns channel
name and units in attributes
}
\examples{
d = read.AR_file(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
attr(d, 'channel')
}
\author{
thomasgredig
}
