% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_file.v2}
\alias{read.AR_file.v2}
\title{loads Asylum Research Igor AFM wavefile}
\usage{
read.AR_file.v2(filename)
}
\arguments{
\item{h1}{filename including path}
}
\value{
\code{AFMdata} object
}
\description{
loads all images contained in Igor wavefile and
returns an AFMdata object
}
