% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_header}
\alias{read.AR_header}
\title{loads Asylum Research Igor Wave AFM files}
\usage{
read.AR_header(filename, no = 1)
}
\arguments{
\item{filename}{filename including path}

\item{no}{number of the channel}
}
\value{
image with attributes
}
\description{
loads Asylum Research Igor Wave AFM files
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
d = read.AR_header(filename)
}
