% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_header.v2}
\alias{read.AR_header.v2}
\title{loads Asylum Research Igor Wave AFM header information}
\usage{
read.AR_header.v2(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
header information
}
\description{
loads Asylum Research Igor Wave AFM header information
}
\examples{
filename = system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM")
d = read.AR_header.v2(filename)
}
