% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.functions.R
\name{read.IBW_file}
\alias{read.IBW_file}
\title{DEPRECATED: use read.AR_file()
loads Asylum Research Igor Wave AFM files}
\usage{
read.IBW_file(filename, no = 1)
}
\arguments{
\item{filename}{filename including path}

\item{no}{number of the channel}
}
\value{
image with attributes
}
\description{
DEPRECATED: use read.AR_file()
loads Asylum Research Igor Wave AFM files
}
\examples{
d = read.IBW_file(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
}
