% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Nanoscope_file.R
\name{read.Nanoscope_BIN}
\alias{read.Nanoscope_BIN}
\title{loads the binary data from the Veeco file}
\usage{
read.Nanoscope_BIN(filename, no = 1)
}
\description{
loads the binary data from the Veeco file
}
