% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Nanoscope_file.R
\name{read.Nanoscope_file}
\alias{read.Nanoscope_file}
\title{loads images of AFM Nanoscope file (use Nanoscope.loadImage whenever)}
\usage{
read.Nanoscope_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list with header, file ID, and images
}
\description{
loads images of AFM Nanoscope file (use Nanoscope.loadImage whenever)
}
\examples{
filename = dir(pattern='000$', recursive=TRUE)[1]
d = read.Nanoscope_file(filename)
}
