% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Nanoscope_file_scaled.R
\name{read.Nanoscope_file_scaled}
\alias{read.Nanoscope_file_scaled}
\title{Returns NanoScope AFM image with Scaling}
\usage{
read.Nanoscope_file_scaled(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with x,y,z data in units of nm
}
\description{
Returns NanoScope AFM image with Scaling
}
\examples{
filename = dir(pattern='000$', recursive=TRUE)[1]
h = read.Nanoscope_file_scaled(filename)
ggplot() + geom_raster(data = d , aes(x = x, y = y, fill = z)) +
coord_equal(expand=FALSE) +
xlab('x (nm)') +
ylab('y (nm)') +
scale_fill_continuous(name='z (nm)')
}
