% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Nanoscope_header.R
\name{read.Nanoscope_header}
\alias{read.Nanoscope_header}
\title{Returns the header of a NanoScope AFM file}
\usage{
read.Nanoscope_header(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with name / value columns
}
\description{
Returns the header of a NanoScope AFM file
}
\examples{
filename = dir(pattern='000$', recursive=TRUE)[1]
h = read.Nanoscope_header(filename)
}
