% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{read.NanoSurf_file}
\alias{read.NanoSurf_file}
\title{loads images of AFM NID file (use NID.loadImage whenever)}
\usage{
read.NanoSurf_file(filename, imageNo = 1)
}
\arguments{
\item{filename}{filename including path}

\item{imageNo}{number of the image file}
}
\value{
AFM image
}
\description{
loads images of AFM NID file (use NID.loadImage whenever)
}
\examples{
filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
d = read.NanoSurf_file(filename)
}
