% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{read.NanoSurf_header}
\alias{read.NanoSurf_header}
\title{loads header of NanoSurf AFM NID file and returns
parameters for particular image}
\usage{
read.NanoSurf_header(filename, imageNo = 1)
}
\arguments{
\item{filename}{filename including path}

\item{imageNo}{image number to get data on}
}
\value{
list
}
\description{
loads header of NanoSurf AFM NID file and returns
parameters for particular image
}
\examples{
filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
read.NanoSurf_header(filename, 1)
}
