% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{read.NanoSurf_header.v2}
\alias{read.NanoSurf_header.v2}
\title{loads header of NanoSurf AFM NID file and returns
parameters for particular image}
\usage{
read.NanoSurf_header.v2(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list
}
\description{
loads header of NanoSurf AFM NID file and returns
parameters for particular image
}
\examples{
filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
h = read.NanoSurf_header.v2(filename)
}
