% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.functions.R
\name{read.NID_file}
\alias{read.NID_file}
\title{DEPRECATED: use AFM.read()
loads images of AFM NID file (use NID.loadImage whenever)}
\usage{
read.NID_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list with header, file ID, and images
}
\description{
DEPRECATED: use AFM.read()
loads images of AFM NID file (use NID.loadImage whenever)
}
\examples{
d = read.NID_file(system.file("extdata","NanoSurf_20160301.nid",package="nanoscopeAFM"))
}
