% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{read.NID_headerItems}
\alias{read.NID_headerItems}
\title{loads header of NID file and separates into items
first item in list has the titles of the others}
\usage{
read.NID_headerItems(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list
}
\description{
loads header of NID file and separates into items
first item in list has the titles of the others
}
\examples{
filename = system.file("extdata", "NanoSurf_20160301.nid",package="nanoscopeAFM")
read.NID_headerItems(filename)
}
