% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NID_sweepFile.R
\name{read.NID_sweepFile}
\alias{read.NID_sweepFile}
\title{loads frequency sweep data from NID file}
\usage{
read.NID_sweepFile(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data.frame with data frames that have freq vs. ampl.
}
\description{
loads frequency sweep data from NID file
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
d = read.NID_sweepFile(filename)
}
