% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Park_file.R
\name{read.Park_header.v2}
\alias{read.Park_header.v2}
\title{loads Park Image header}
\usage{
read.Park_header.v2(filename)
}
\arguments{
\item{filename}{Park AFM filename including path}
}
\value{
AFMinfo data
}
\description{
loads Park Image header
}
\examples{
filename = system.file("extdata","Park_20210916_034.tiff",package="nanoscopeAFM")
h = read.Park_header.v2(filename)
}
\author{
Thomas Gredig
}
