% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{summary.AFMdata}
\alias{summary.AFMdata}
\title{summary of AFMdata object}
\usage{
\method{summary}{AFMdata}(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
summary of AFMdata object
}
\description{
summary of AFMdata object
}
\examples{
d = AFM.import(AFM.getSampleImages(type='ibw')[1])
summary(d)
}
\author{
Thomas Gredig
}
