% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMinfo.R
\name{summary.AFMinfo}
\alias{summary.AFMinfo}
\title{summary for AFMinfo object}
\usage{
\method{summary}{AFMinfo}(obj)
}
\arguments{
\item{obj}{AFMinfo object}
}
\value{
quick summary
}
\description{
summary for AFMinfo object
}
\examples{
h1 = AFMinfo(AFM.getSampleImages(type='ibw')[1])
summary(h1)
}
\author{
Thomas Gredig
}
