% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.math.R
\name{summary.AFMmath}
\alias{summary.AFMmath}
\title{Summary of computed AFM image parameters}
\usage{
\method{summary}{AFMmath}(obj)
}
\arguments{
\item{obj}{AFMmath object}
}
\value{
prints a summary
}
\description{
Summary of computed AFM image parameters
}
\examples{
filename = system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM")
summary(AFM.math.params(AFM.import(filename)))
}
