% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flatten}
\alias{AFM.flatten}
\title{flattens an AFM image using a plane fit}
\usage{
AFM.flatten(d)
}
\arguments{
\item{d}{data.frame with AFM image}
}
\value{
flattened matrix with AFM image
}
\description{
flattens an AFM image using a plane fit
}
\examples{
filename = dir(pattern='\\\\d$', recursive=TRUE)[1]
d = AFM.read(filename)
d$z.flatten = AFM.flatten(d)
}
