% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flatten}
\alias{AFM.flatten}
\title{flattens an AFM image using a plane fit
this flattens only z, not z.nm, so scaling
factor would need to be applied for z.nm}
\usage{
AFM.flatten(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
flattened matrix with AFM image
}
\description{
flattens an AFM image using a plane fit
this flattens only z, not z.nm, so scaling
factor would need to be applied for z.nm
}
\examples{
d = AFM.read(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
d2 = AFM.flatten(d)
}
\author{
thomasgredig
}
