% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.flatten.R
\name{AFM.flatten}
\alias{AFM.flatten}
\title{flattens an AFM image using a plane fit}
\usage{
AFM.flatten(obj, no = 1)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{Image number}
}
\value{
flattened matrix with AFM image
}
\description{
this flattens only z of one of the images
using a plane fit
}
\examples{
d = AFM.read(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
d2 = AFM.flatten(d)
plot(d2)
}
\author{
thomasgredig
}
