% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.getSampleImages.R
\name{AFM.getSampleImages}
\alias{AFM.getSampleImages}
\title{sample AFM images}
\usage{
AFM.getSampleImages()
}
\value{
vector with path/filename to AFM sample images
}
\description{
returns sample AFM images from library
}
\examples{
file.list = AFM.getSampleImages()
print(paste("Found",length(file.list),"sample files."))
}
\author{
Thomas Gredig
}
