% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.heightHeightCorrelationFunction.R
\name{AFM.hhcf}
\alias{AFM.hhcf}
\title{Height-Height Correlation Function}
\usage{
AFM.hhcf(
  obj,
  no = 1,
  degRes = 100,
  numIterations = 1000,
  addFit = TRUE,
  dataOnly = FALSE,
  r.percentage = 80,
  verbose = FALSE
)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{Channel number}

\item{degRes}{resolution of angle, the higher the more, should be >100, 1000 is also good}

\item{numIterations}{Number of iterations (must be > 1000), but 1e6 recommended}

\item{addFit}{if \code{TRUE} a fit is added to the data}

\item{dataOnly}{if \code{TRUE} only return data frame, otherwise returns a graph}

\item{r.percentage}{a number from 10 to 100 representing the distance to compute, since the image is
square, there are not as many points that are separate by the full lenght, 80 is a good value, if there
is no fit, the value can be reduced to 70 or 60.}

\item{verbose}{output time if \code{TRUE}}
}
\value{
graph or data frame with g(r) and $num indicating number of computations used for r
}
\description{
Computes g(r) correlation function for an AFM
image with height information. The height information
must be encoded in a BMP file with a linear scale
Publication: http://iopscience.iop.org/article/10.1088/1742-6596/417/1/012069
Title: Height-Height Correlation Function to Determine Grain
       Size in Iron Phthalocyanine Thin Films
Authors: Thomas Gredig, Evan A. Silverstein, Matthew P Byrne
Journal: J of Phys: Conf. Ser. Vol 417, p. 012069 (2013).
}
\examples{
filename = AFM.getSampleImages(type='tiff')
a = AFM.import(filename)
a = AFM.flatten(a)
r = AFM.hhcf(a, dataOnly = TRUE)
head(r)

}
\author{
Thomas Gredig
}
