% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.histogram}
\alias{AFM.histogram}
\title{graph a histogram for the AFM image}
\usage{
AFM.histogram(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
draws a ggplot graph
}
\description{
graph a histogram for the AFM image
}
\examples{
library(ggplot2)
d = AFM.import(AFM.getSampleImages(type='ibw')[1])
AFM.histogram(d)
}
\author{
Thomas Gredig
}
