% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.histogram}
\alias{AFM.histogram}
\title{graph a histogram of the height distribution for an AFMdata object}
\usage{
AFM.histogram(AFMdata)
}
\arguments{
\item{AFMdata}{AFMdata object}
}
\value{
ggplot graph
}
\description{
graph a histogram of the height distribution for an AFMdata object
}
\examples{
library(ggplot2)
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
AFM.histogram(d)
}
\author{
thomasgredig
}
