% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.histogram}
\alias{AFM.histogram}
\title{graph a histogram for the AFM image}
\usage{
AFM.histogram(obj, no = 1, dataOnly = FALSE)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number of the image}

\item{dataOnly}{if \code{TRUE} a data frame with the histogram data is returned}
}
\value{
draws a ggplot graph
}
\description{
graph a histogram for the AFM image
}
\examples{
library(ggplot2)
d = AFM.import(AFM.getSampleImages(type='ibw')[1])
AFM.histogram(d)
head(AFM.histogram(d, dataOnly=TRUE),n=20)
}
\author{
Thomas Gredig
}
