% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.import}
\alias{AFM.import}
\title{imports AFM file}
\usage{
AFM.import(filename)
}
\arguments{
\item{filename}{name of AFM filename}
}
\value{
AFMdata object
}
\description{
imports AFM file
}
\examples{
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
}
\author{
Thomas Gredig
}
