% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.import}
\alias{AFM.import}
\title{Imports AFM file}
\usage{
AFM.import(filename)
}
\arguments{
\item{filename}{name of AFM filename}
}
\value{
AFMdata object
}
\description{
Use this function to create an AFMdata object from the filename
}
\examples{
library(ggplot2)
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
summary(d)
plot(d)
}
\author{
Thomas Gredig
}
