% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.raster}
\alias{AFM.raster}
\title{Raster data frame}
\usage{
AFM.raster(obj, no = 1)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{channel number}
}
\value{
data.frame with ($x, $y, $z) raster image; ($x,$y) in units of nm
}
\description{
data frame has  ($x, $y, $z) in units for particular channel, ($x, $y) are
always in units of nanometer
}
\examples{
afmd = AFM.import(AFM.getSampleImages(type='ibw')[1])
d = AFM.raster(afmd, 1)
head(d)
}
\author{
Thomas Gredig
}
