% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.raster}
\alias{AFM.raster}
\title{returns raster data frame}
\usage{
AFM.raster(obj, no = 1)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{image number}
}
\value{
data.frame with ($x, $y, $z) raster image; ($x,$y) in units of nm
}
\description{
data frame has  ($x, $y, $z) in units of particular channel, ($x, $y) are
always in units of nanometer
}
\author{
Thomas Gredig
}
