% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{AFM.raster}
\alias{AFM.raster}
\title{returns data frame with ($x.nm, $y.nm, $z.nm) in nanometers}
\usage{
AFM.raster(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
data.frame with  ($x.nm, $y.nm, $z.nm)
}
\description{
returns data frame with ($x.nm, $y.nm, $z.nm) in nanometers
}
\author{
thomasgredig
}
