% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.read.R
\name{AFM.read}
\alias{AFM.read}
\title{loads AFM image file}
\usage{
AFM.read(filename, no = 1)
}
\arguments{
\item{filename}{filename of (Veeco, Park, AR, NanoSurf) AFM image including path}

\item{no}{channel number (for Veeco, NanoSurf, AR)}
}
\value{
AFM image with attributes
}
\description{
\code{AFM.read} reads data files from different instruments
and stores 6 rows (x,y,z) and converted units, when possible
(xc,yc,zc), use the attributes to find the units and
channel name
}
\examples{
filename = system.file("extdata","Park_20210916_034.tiff",package="nanoscopeAFM")
d = AFM.read(filename)
}
\author{
thomasgredig
}
