% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMinfo.R
\name{AFMinfo.item}
\alias{AFMinfo.item}
\title{returns an item from the AFMinfo object}
\usage{
AFMinfo.item(obj, itemName)
}
\arguments{
\item{obj}{AFMinfo object}

\item{itemName}{name to retrieve (ScanRate, ScanAngle, ...)}
}
\value{
string or number
}
\description{
returns an item from the AFMinfo object
}
\examples{
filename = system.file("extdata","Veeco_20160622.003",package="nanoscopeAFM")
h = AFMinfo(filename)
AFMinfo.item(h,"Description")
AFMinfo.item(h,"Operating mode")
}
\author{
Thomas Gredig
}
