% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFM.info.R
\name{AFMinfo}
\alias{AFMinfo}
\title{AFMinfo object with information about the AFM image}
\usage{
AFMinfo(filename)
}
\arguments{
\item{filename}{filename of (Veeco, Park, AR, NanoSurf) AFM image including path}
}
\value{
AFMinfo object
}
\description{
AFMinfo object with information about the AFM image
}
\author{
Thomas Gredig
}
