% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.NID_file.R
\name{check.NID_file}
\alias{check.NID_file}
\title{checks AFM NID file length}
\usage{
check.NID_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
mismatch in image size + header with file size (should be 0)
}
\description{
checks AFM NID file length
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
d = check.NID_file(filename)
}
