% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NanoSurf_file.R
\name{get.NID_imageInfo}
\alias{get.NID_imageInfo}
\title{returns number of lines for each image}
\usage{
get.NID_imageInfo(header.string)
}
\arguments{
\item{header.string}{header string of NID file (use read.NID_header)}
}
\value{
vector with line numbers for each image
}
\description{
returns number of lines for each image
}
\examples{
hdr = read.NID_header(system.file("extdata","Veeco_20160622.003",package="nanoscopeAFM"))
get.NID_imageInfo(hdr[[2]])
}
