% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.NID_imageInfo.R
\name{get.NID_imageInfo}
\alias{get.NID_imageInfo}
\title{returns number of lines for each image}
\usage{
get.NID_imageInfo(header.string)
}
\arguments{
\item{header.string}{header string of NID file (use read.NID_header)}
}
\value{
vector with line numbers for each image
}
\description{
returns number of lines for each image
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
hdr = read.NID_header(filename)
get.NID_imageInfo(hdr[[2]])
}
