% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{initialize,AFMdata-method}
\alias{initialize,AFMdata-method}
\title{Constructor method of AFMImage Class.}
\usage{
\S4method{initialize}{AFMdata}(
  .Object,
  data,
  x.max.nm,
  y.max.nm,
  x.pixels,
  y.pixels,
  z.conv,
  z.units,
  channel,
  instrument,
  fullfilename
)
}
\arguments{
\item{.Object}{an AFMdata object}
}
\description{
Constructor method of AFMImage Class.
}
\section{Slots}{

\describe{
\item{\code{data}}{($x,$y,$z): a data.frame storing the coordinates of the sample and the measured heights}

\item{\code{x.max.nm}}{maximum width in units of nm}

\item{\code{y.max.nm}}{maximum height in units of nm}

\item{\code{x.pixels}}{number of pixels in x-direction}

\item{\code{y.pixels}}{number of pixels in y-direction}

\item{\code{z.conv}}{conversion factor for z to convert to $units}

\item{\code{z.units}}{units for z (deg, m)}

\item{\code{channel}}{name of channel}

\item{\code{instrument}}{name of instrument (Park, AR, NanoSurf, Veeco)}

\item{\code{fullfilename}}{name of filename}
}}

