% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.functions.R
\name{NID.loadImage}
\alias{NID.loadImage}
\title{DEPRACATED: returns the scaled image + flattend image}
\usage{
NID.loadImage(filename, imageNo = 1)
}
\arguments{
\item{filename}{filename including path}

\item{imageNo}{a number of the image, 1,2,3,4}
}
\value{
data.frame with four components, x and y-axes
        and the z-axis (original) and flattened
}
\description{
DEPRACATED: returns the scaled image + flattend image
}
\examples{
filename = system.file("extdata","NanoSurf_20160301.nid",package="nanoscopeAFM")
d = NID.loadImage(filename,1)
ggplot(d, aes(x,y,fill=z.flatten)) + geom_raster()
}
