% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NID.loadSweep.R
\name{NID.loadSweep}
\alias{NID.loadSweep}
\title{loads frequency sweep data from NID file}
\usage{
NID.loadSweep(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data.frame with data frames that have freq vs. ampl.
}
\description{
loads frequency sweep data from NID file
}
\examples{
d = NID.loadSweep(system.file("extdata","NanoSurf_20160301.nid",package="nanoscopeAFM"))
}
