% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{plot.AFMdata}
\alias{plot.AFMdata}
\title{graph of AFMdata object}
\usage{
\method{plot}{AFMdata}(obj, no = 1, mpt = NA, ...)
}
\arguments{
\item{obj}{AFMdata object}

\item{no}{number of the image}

\item{mpt}{midpoint for coloring}
}
\value{
ggplot graph
}
\description{
graph of AFMdata object
}
\examples{
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
plot(d)
}
\author{
Thomas Gredig
}
