% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AFM_file.R
\name{read.AFM_file}
\alias{read.AFM_file}
\title{loads AFM image file}
\usage{
read.AFM_file(filename, no = 1)
}
\arguments{
\item{filename}{filename of AFM image raw data including path}

\item{no}{number of the channel}
}
\value{
AFM image with attributes
}
\description{
loads AFM image file
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
d = read.AFM_file(filename)
}
