% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_eofHeader.V2}
\alias{read.AR_eofHeader.V2}
\title{returns names of AR channel names}
\usage{
read.AR_eofHeader.V2(wavefile, Verbose = FALSE)
}
\arguments{
\item{wavefile}{filename including path}

\item{Verbose}{if true, returns additional information}
}
\value{
list of channels and additional information
}
\description{
returns names of AR channel names
}
\examples{
filename = dir(pattern='ibw$', recursive=TRUE)[1]
afmHeadInfo = read.AR_eofHeader.V2(filename)
}
