% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.AR_file.R
\name{read.AR_header}
\alias{read.AR_header}
\title{loads AR header}
\usage{
read.AR_header(filename, no = 1)
}
\arguments{
\item{filename}{filename including path}

\item{no}{number of the channel}
}
\value{
image with attributes
}
\description{
use this function to read the header information
}
\examples{
d = read.AR_header(system.file("AR_20211011.ibw",package="nanoscopeAFM"))
}
