% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Nanoscope_file.R
\name{read.Nanoscope_file}
\alias{read.Nanoscope_file}
\title{Returns Veeco NanoScope AFM image with Scaling}
\usage{
read.Nanoscope_file(filename, no = 1)
}
\arguments{
\item{filename}{filename including path}

\item{no}{image number}
}
\value{
image
}
\description{
Returns Veeco NanoScope AFM image with Scaling
}
\examples{
filename = dir(pattern='\\\\d$', recursive=TRUE)[1]
h = read.Nanoscope_file(filename)
}
