% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NID_file.R
\name{read.NID_file}
\alias{read.NID_file}
\title{loads images of AFM NID file (use NID.loadImage whenever)}
\usage{
read.NID_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list with header, file ID, and images
}
\description{
loads images of AFM NID file (use NID.loadImage whenever)
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
d = read.NID_file(filename)
}
