% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NID_header.R
\name{read.NID_file}
\alias{read.NID_file}
\title{loads images of AFM NID file}
\usage{
read.NID_file(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
images
}
\description{
loads images of AFM NID file
}
\examples{
d = read.NID_file('example.nid')
}
