% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.NID_headerItems.R
\name{read.NID_headerItems}
\alias{read.NID_headerItems}
\title{loads header of NID file and separates into items}
\usage{
read.NID_headerItems(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list
}
\description{
loads header of NID file and separates into items
}
\examples{
filename = dir(pattern='nid$', recursive=TRUE)[1]
read.NID_headerItems(filename)
}
