% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.Park_file.R
\name{read.Park_file}
\alias{read.Park_file}
\title{loads images}
\usage{
read.Park_file(filename)
}
\arguments{
\item{filename}{Park AFM filename including path}
}
\value{
image with attributes
}
\description{
loads images
}
\examples{
filename = system.file("extdata","Park_20210916_034.tiff",package="nanoscopeAFM")
d = read.Park_file(filename)
}
\author{
thomasgredig
}
