% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFMdata.R
\name{summary.AFMdata}
\alias{summary.AFMdata}
\title{summary of AFMdata object}
\usage{
\method{summary}{AFMdata}(obj)
}
\arguments{
\item{obj}{AFMdata object}
}
\value{
summary of AFMdata object
}
\description{
summary of AFMdata object
}
\examples{
d = AFM.import(system.file("extdata","AR_20211011.ibw",package="nanoscopeAFM"))
summary(d)
}
\author{
Thomas Gredig
}
